Fixed bugs in sharpness and highlights tests
The failing tests for `changeSharpness` were caused by the original arrays being modified in place, and then resulting always equal to the edited copies. By making an actual copy of the original array through slicing, the tests now work properly.

The failing test for `changeHighlights` was caused by a typo - the `changeShadows` function was called instead of `changeHighlights` in one of the tests.